+++ src/eval/evaluator_par.hpp
@@ -10,6 +10,8 @@ class PartialEvaluator {
 
   void initZeros(size_t from, size_t to);
 
+  bool checkValue(int64_t cell, int64_t expected_value) const;
+
   bool doPartialEval(Program& p, size_t op_index);
 
   Operand resolveOperand(const Operand& op) const;